+++ sdk/cost-management/arm-costmanagement/src/api/forecast/options.ts
@@ -8,7 +8,6 @@ export interface ForecastExternalCloudProviderUsageOptionalParams extends Operat
   /** May be used to filter forecasts by properties/usageDate (Utc time), properties/chargeType or properties/grain. The filter supports 'eq', 'lt', 'gt', 'le', 'ge', and 'and'. It does not currently support 'ne', 'or', or 'not'. */
   filter?: string;
 }
-
 /** Optional parameters. */
 export interface ForecastUsageOptionalParams extends OperationOptions {
   /** May be used to filter forecasts by properties/usageDate (Utc time), properties/chargeType or properties/grain. The filter supports 'eq', 'lt', 'gt', 'le', 'ge', and 'and'. It does not currently support 'ne', 'or', or 'not'. */