+++ sdk/cost-management/arm-costmanagement/src/api/markupRules/index.ts
@@ -1,0 +1,10 @@
+// Copyright (c) Microsoft Corporation.
+// Licensed under the MIT License.
+
+export { list, $delete, createOrUpdate, get } from "./operations.js";
+export type {
+  MarkupRulesListOptionalParams,
+  MarkupRulesDeleteOptionalParams,
+  MarkupRulesCreateOrUpdateOptionalParams,
+  MarkupRulesGetOptionalParams,
+} from "./options.js";