+++ sdk/cost-management/arm-costmanagement/CHANGELOG.md
@@ -1,4 +1,22 @@
 # Release History
+
+## 1.1.0 (2026-08-12)
+
+### Features Added
+  - Added operation group MarkupRulesOperations
+  - Added Interface CustomerMetadata
+  - Added Interface ManagementGroupScopeBenefitRecommendationProperties
+  - Added Interface MarkupRule
+  - Added Interface MarkupRuleProperties
+  - Added Interface MarkupRulesCreateOrUpdateOptionalParams
+  - Added Interface MarkupRulesDeleteOptionalParams
+  - Added Interface MarkupRulesGetOptionalParams
+  - Added Interface MarkupRulesListOptionalParams
+  - Enum KnownScope has a new value ManagementGroup
+  - Enum KnownVersions has a new value V20260601
+  - Operation ForecastOperations.usage has a new signature
+  - Operation QueryOperations.usage has a new signature
+
     
 ## 1.0.0 (2026-07-16)